requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.101.1
filelock==3.12.2
firebase-admin==6.2.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
konlpy==0.6.0
lightning-utilities==0.9.0
lxml==4.9.3
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
proto-plus==1.22.3
protobuf==4.24.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.0
pydantic_core==2.6.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.0.7
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rsa==4.9
safetensors==0.3.2
schedule==1.2.0
scikit-learn==1.3.0
scipy==1.11.1
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchmetrics==1.0.3
tqdm==4.66.1
transformers==4.31.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
wandb==0.15.8
yarl==1.9.2
